fix(plugin-blanks): different approach to allowing blanks in tables only within blanks exercise
