+++ packages/editor/src/plugins/serlo-table/editor.tsx
@@ -1,5 +1,4 @@
 import { EditorTooltip } from '@editor/editor-ui/editor-tooltip'
-import { TextEditorFormattingOption } from '@editor/editor-ui/plugin-toolbar/text-controls/types'
 import {
   store,
   selectFocused,
@@ -23,21 +22,6 @@ import { SerloTableToolbar } from './toolbar'
 import { getTableType } from './utils/get-table-type'
 import { TextEditorConfig } from '../text'
 import { instanceStateStore } from '../text/utils/instance-state-store'
-
-const headerTextFormattingOptions = [
-  TextEditorFormattingOption.code,
-  TextEditorFormattingOption.links,
-  TextEditorFormattingOption.math,
-]
-const cellTextFormattingOptions = [
-  TextEditorFormattingOption.code,
-  TextEditorFormattingOption.colors,
-  TextEditorFormattingOption.links,
-  TextEditorFormattingOption.lists,
-  TextEditorFormattingOption.math,
-  TextEditorFormattingOption.richTextBold,
-  TextEditorFormattingOption.richTextItalic,
-]
 
 const newCell = { content: { plugin: EditorPluginType.Text } }
 
@@ -95,14 +79,6 @@ export function SerloTableEditor(props: SerloTableProps) {
             store.getState(),
             cell.content.id
           )
-          const formattingOptions = [
-            ...(isHead
-              ? headerTextFormattingOptions
-              : cellTextFormattingOptions),
-            ...(props.config.allowBlanks
-              ? [TextEditorFormattingOption.textBlank]
-              : []),
-          ]
 
           const onKeyUpHandler = (e: KeyboardEvent<HTMLDivElement>) => {
             // hack: redraw when isEmpty changes. (onKeyUp bc. keyDown is captured for some keys)
@@ -137,7 +113,9 @@ export function SerloTableEditor(props: SerloTableProps) {
                 config: {
                   isInlineChildEditor: true,
                   placeholder: '',
-                  formattingOptions,
+                  formattingOptions: isHead
+                    ? props.config.headerTextFormattingOptions
+                    : props.config.cellTextFormattingOptions,
                 } as TextEditorConfig,
               })}
               {props.config.allowImageInTableCells && !areImagesDisabled ? (
